backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymysql==1.1.0
pymongo==4.6.1
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
marshmallow==3.20.1
flask-limiter==3.5.0
python-dateutil==2.8.2
cryptography==41.0.7